introduce interface for sources
Re-arranged the structure by introducing a "source" interface that
can be either a program or a whois server.
The problem was that the automatic fallback mechanism would not
differentiate between a program and a hostname (and the thing is not
trivial to automate), so the choice is to leave the client to decide
what to use, as it has more clues than just a string (see flags in
main).